refactor(runtime): extract MCP protocol types into mcp_types module
Move all JSON-RPC and MCP protocol type definitions (JsonRpcRequest,
McpTool, McpServerManagerError, etc.) from mcp_stdio.rs into a
dedicated mcp_types.rs module. mcp_stdio.rs retains server lifecycle
management, stdio process I/O, and tests.
