+++ crates/runtime/src/mcp_stdio.rs
@@ -3,7 +3,7 @@ use std::io;
 use std::process::Stdio;
 
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::Value as JsonValue;
 use tokio::io::{AsyncBufReadExt, AsyncReadExt, AsyncWriteExt, BufReader};
 use tokio::process::{Child, ChildStdin, ChildStdout, Command};
@@ -11,279 +11,12 @@ use tokio::process::{Child, ChildStdin, ChildStdout, Command};
 use crate::config::{McpTransport, RuntimeConfig, ScopedMcpServerConfig};
 use crate::mcp::mcp_tool_name;
 use crate::mcp_client::{McpClientBootstrap, McpClientTransport, McpStdioTransport};
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(untagged)]
-pub enum JsonRpcId {
-    Number(u64),
-    String(String),
-    Null,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct JsonRpcRequest<T = JsonValue> {
-    pub jsonrpc: String,
-    pub id: JsonRpcId,
-    pub method: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub params: Option<T>,
-}
-
-impl<T> JsonRpcRequest<T> {
-    #[must_use]
-    pub fn new(id: JsonRpcId, method: impl Into<String>, params: Option<T>) -> Self {
-        Self {
-            jsonrpc: "2.0".to_string(),
-            id,
-            method: method.into(),
-            params,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct JsonRpcError {
-    pub code: i64,
-    pub message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct JsonRpcResponse<T = JsonValue> {
-    pub jsonrpc: String,
-    pub id: JsonRpcId,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub result: Option<T>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<JsonRpcError>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpInitializeParams {
-    pub protocol_version: String,
-    pub capabilities: JsonValue,
-    pub client_info: McpInitializeClientInfo,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpInitializeClientInfo {
-    pub name: String,
-    pub version: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpInitializeResult {
-    pub protocol_version: String,
-    pub capabilities: JsonValue,
-    pub server_info: McpInitializeServerInfo,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpInitializeServerInfo {
-    pub name: String,
-    pub version: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpListToolsParams {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct McpTool {
-    pub name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(rename = "inputSchema", skip_serializing_if = "Option::is_none")]
-    pub input_schema: Option<JsonValue>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub annotations: Option<JsonValue>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpListToolsResult {
-    pub tools: Vec<McpTool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub next_cursor: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpToolCallParams {
-    pub name: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub arguments: Option<JsonValue>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct McpToolCallContent {
-    #[serde(rename = "type")]
-    pub kind: String,
-    #[serde(flatten)]
-    pub data: BTreeMap<String, JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpToolCallResult {
-    #[serde(default)]
-    pub content: Vec<McpToolCallContent>,
-    #[serde(default)]
-    pub structured_content: Option<JsonValue>,
-    #[serde(default)]
-    pub is_error: Option<bool>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpListResourcesParams {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub cursor: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct McpResource {
-    pub uri: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(rename = "mimeType", skip_serializing_if = "Option::is_none")]
-    pub mime_type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub annotations: Option<JsonValue>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpListResourcesResult {
-    pub resources: Vec<McpResource>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub next_cursor: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[serde(rename_all = "camelCase")]
-pub struct McpReadResourceParams {
-    pub uri: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct McpResourceContents {
-    pub uri: String,
-    #[serde(rename = "mimeType", skip_serializing_if = "Option::is_none")]
-    pub mime_type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub text: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub blob: Option<String>,
-    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
-    pub meta: Option<JsonValue>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct McpReadResourceResult {
-    pub contents: Vec<McpResourceContents>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct ManagedMcpTool {
-    pub server_name: String,
-    pub qualified_name: String,
-    pub raw_name: String,
-    pub tool: McpTool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct UnsupportedMcpServer {
-    pub server_name: String,
-    pub transport: McpTransport,
-    pub reason: String,
-}
-
-#[derive(Debug)]
-pub enum McpServerManagerError {
-    Io(io::Error),
-    JsonRpc {
-        server_name: String,
-        method: &'static str,
-        error: JsonRpcError,
-    },
-    InvalidResponse {
-        server_name: String,
-        method: &'static str,
-        details: String,
-    },
-    UnknownTool {
-        qualified_name: String,
-    },
-    UnknownServer {
-        server_name: String,
-    },
-}
-
-impl std::fmt::Display for McpServerManagerError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Io(error) => write!(f, "{error}"),
-            Self::JsonRpc {
-                server_name,
-                method,
-                error,
-            } => write!(
-                f,
-                "MCP server `{server_name}` returned JSON-RPC error for {method}: {} ({})",
-                error.message, error.code
-            ),
-            Self::InvalidResponse {
-                server_name,
-                method,
-                details,
-            } => write!(
-                f,
-                "MCP server `{server_name}` returned invalid response for {method}: {details}"
-            ),
-            Self::UnknownTool { qualified_name } => {
-                write!(f, "unknown MCP tool `{qualified_name}`")
-            }
-            Self::UnknownServer { server_name } => write!(f, "unknown MCP server `{server_name}`"),
-        }
-    }
-}
-
-impl std::error::Error for McpServerManagerError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            Self::Io(error) => Some(error),
-            Self::JsonRpc { .. }
-            | Self::InvalidResponse { .. }
-            | Self::UnknownTool { .. }
-            | Self::UnknownServer { .. } => None,
-        }
-    }
-}
-
-impl From<io::Error> for McpServerManagerError {
-    fn from(value: io::Error) -> Self {
-        Self::Io(value)
-    }
-}
+use crate::mcp_types::{
+    JsonRpcId, JsonRpcRequest, JsonRpcResponse, ManagedMcpTool, McpInitializeClientInfo,
+    McpInitializeParams, McpInitializeResult, McpListResourcesParams, McpListResourcesResult,
+    McpListToolsParams, McpListToolsResult, McpReadResourceParams, McpReadResourceResult,
+    McpServerManagerError, McpToolCallParams, McpToolCallResult, UnsupportedMcpServer,
+};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 struct ToolRoute {
